requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.2rc1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
